Add optional gesture conformance
